Add delete board feature

+++ backend/src/main/java/com/it4us/todoapp/service/BoardServiceImpl.java
@@ -147,8 +147,11 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public void deleteBoard(Long id) {
+    public void deleteBoard(Long id, String username) {
         Board board=boardRepository.findById(id).orElseThrow(()->new NotFoundException("board not found"));
+        if(boardRepository.isBoardBelongedUser(id,username)<1){
+            throw new BoardBelongAnotherUserException("Board belongs to another user");
+        }
         boardRepository.delete(board);
     }
